src/ponderada02/src/requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
